Add turn logic
The turn is initialized as 1. The game.playTurn() function calls the
gameBoard.mark() function to attempt to mark a square.
If the square is occupied, mark returns, and game.playTurn() returns
before increasing the turn number.
If the square was filled, mark returns true and then playTurn does
increase the turn number.