Check if there is a winner / Print board to console
Starting on the 5th turn, game.playTurn() will use
gameBoard.checkWinner() to verify if there is a winner. If it returns
true then game.playTurn() announces that the current player (and
therefore the latest player to make a move) won.

game.printBoard() was added to represent the board in the console for
easy debugging. It prints a string representing the squares. It creates
a newline for each row.