Add player names and preliminary rename UI
Player objects now have a setName and getName function. These are not
directly called by the UI, but rather by the drawController module.

When the player's name is clicked, drawController.rename() prompts the
player for a new name. Data validation occurs at this stage.
If the name is valid (1-8 chars) then the function is called on the
appropriate player, determined with the data-player attribute of the
player's name DOM element.

This works, but the ability to rename will be put in a pencil icon in
the future so it is immediately obvious to the player.

Also styled the reset button slightly.